requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-chroma
chromadb
pypdf
python-dotenv
